% Auto-generated: do not edit by hand
\name{''BreadcrumbsDashUiComponents}

\alias{''BreadcrumbsDashUiComponents}

\title{BreadcrumbsDashUiComponents component}

\description{

}

\usage{
''BreadcrumbsDashUiComponents(id=NULL, model=NULL, home=NULL, style=NULL, className=NULL)
}

\arguments{
\item{id}{Character. }

\item{model}{Unnamed list. }

\item{home}{Logical | numeric | character | named list | unnamed list. }

\item{style}{Named list. }

\item{className}{Character. }
}
