% Auto-generated: do not edit by hand
\name{''CaptchaDashUiComponents}

\alias{''CaptchaDashUiComponents}

\title{CaptchaDashUiComponents component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''CaptchaDashUiComponents(id=NULL, sitekey=NULL, theme=NULL, captchaType=NULL, size=NULL, index=NULL,
language=NULL, onResponse=NULL, onExpire=NULL, responsetext=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{sitekey}{Character. The Sitekey is google recaptcha sitekey}

\item{theme}{Character. }

\item{captchaType}{Character. }

\item{size}{Character. }

\item{index}{Numeric. }

\item{language}{Character. }

\item{onResponse}{}

\item{onExpire}{}

\item{responsetext}{Character. }
}
