% Auto-generated: do not edit by hand
\name{''CurrentPageReport}

\alias{''CurrentPageReport}

\title{CurrentPageReport component}

\description{

}

\usage{
''CurrentPageReport(id=NULL, pageCount=NULL, page=NULL, template=NULL)
}

\arguments{
\item{id}{Character. }

\item{pageCount}{Numeric. }

\item{page}{Numeric. }

\item{template}{Character. }
}
