% Auto-generated: do not edit by hand
\name{''DataScrollerDashUiComponents}

\alias{''DataScrollerDashUiComponents}

\title{DataScrollerDashUiComponents component}

\description{

}

\usage{
''DataScrollerDashUiComponents(id=NULL, value=NULL, rows=NULL, inline=NULL, scrollHeight=NULL, loader=NULL,
buffer=NULL, style=NULL, className=NULL, onLazyLoad=NULL, itemTemplate=NULL,
header=NULL, footer=NULL, lazy=NULL)
}

\arguments{
\item{id}{Character. }

\item{value}{Unnamed list. }

\item{rows}{Numeric. }

\item{inline}{Logical. }

\item{scrollHeight}{Logical | numeric | character | named list | unnamed list. }

\item{loader}{Logical | numeric | character | named list | unnamed list. }

\item{buffer}{Numeric. }

\item{style}{Named list. }

\item{className}{Character. }

\item{onLazyLoad}{}

\item{itemTemplate}{}

\item{header}{Logical | numeric | character | named list | unnamed list. }

\item{footer}{Logical | numeric | character | named list | unnamed list. }

\item{lazy}{Logical. }
}
