% Auto-generated: do not edit by hand
\name{''Geolocator}

\alias{''Geolocator}

\title{Geolocator component}

\description{

}

\usage{
''Geolocator(positionOptions=NULL, watchPosition=NULL, userDecisionTimeout=NULL,
suppressLocationOnMount=NULL, geolocationProvider=NULL,
isOptimisticGeolocationEnabled=NULL, onSuccess=NULL, onError=NULL, errorMsg=NULL)
}

\arguments{
\item{positionOptions}{Named list. }

\item{watchPosition}{Logical. }

\item{userDecisionTimeout}{Numeric. }

\item{suppressLocationOnMount}{Logical. }

\item{geolocationProvider}{Named list. }

\item{isOptimisticGeolocationEnabled}{Logical. }

\item{onSuccess}{}

\item{onError}{}

\item{errorMsg}{Named list. }
}
