% Auto-generated: do not edit by hand
\name{''ListBoxHeader}

\alias{''ListBoxHeader}

\title{ListBoxHeader component}

\description{
Header component of the listbox
}

\usage{
''ListBoxHeader(filter=NULL, disabled=NULL, onFilter=NULL)
}

\arguments{
\item{filter}{Character. When specified, displays a filter input at header.}

\item{disabled}{Logical. When specified, disables the component.}

\item{onFilter}{Callback to invoke when value of filter list changes}
}
