% Auto-generated: do not edit by hand
\name{''ListBoxItem}

\alias{''ListBoxItem}

\title{ListBoxItem component}

\description{
Component to render individual items in a list box
}

\usage{
''ListBoxItem(option=NULL, label=NULL, selected=NULL, tabIndex=NULL, onClick=NULL,
onTouchEnd=NULL, template=NULL)
}

\arguments{
\item{option}{Logical | numeric | character | named list | unnamed list. Listbox option}

\item{label}{Character. Label to display}

\item{selected}{Logical. Boolean. If the option has been selected or not}

\item{tabIndex}{Character. Index of the option}

\item{onClick}{Callback to invoke when the option is clicked.}

\item{onTouchEnd}{Callback to invoke when the option is clicked.}

\item{template}{Template funtion to use}
}
