% Auto-generated: do not edit by hand
\name{''MenuBarComponent}

\alias{''MenuBarComponent}

\title{MenuBarComponent component}

\description{
MenuBarComponent
}

\usage{
''MenuBarComponent(id=NULL, model=NULL, style=NULL, inputPlaceholder=NULL, btnicon=NULL,
btnlabel=NULL, className=NULL, menubarItem=NULL, submenubarItem=NULL,
inputTextBox=NULL, button=NULL, searchQuery=NULL, menubarInput=NULL,
btnClick=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{style}{Named list. Inline style of the component.}

\item{inputPlaceholder}{Character. Input placeholder}

\item{btnicon}{Character. Button Icon}

\item{btnlabel}{Character. Button label}

\item{className}{Character. Style class of the component.}

\item{menubarItem}{Named list. Current Menu object}

\item{submenubarItem}{Named list. Currently selected item}

\item{inputTextBox}{Logical. Show or hide input text box}

\item{button}{Logical. Show or hide button}

\item{searchQuery}{Character. }

\item{menubarInput}{Character. }

\item{btnClick}{Character. }
}
