% Auto-generated: do not edit by hand
\name{''MenubarSub}

\alias{''MenubarSub}

\title{MenubarSub component}

\description{

}

\usage{
''MenubarSub(model=NULL, root=NULL, className=NULL, popup=NULL, onLeafClick=NULL,
onKeyDown=NULL, parentActive=NULL)
}

\arguments{
\item{model}{Logical | numeric | character | named list | unnamed list. }

\item{root}{Logical. }

\item{className}{Character. }

\item{popup}{Logical. }

\item{onLeafClick}{}

\item{onKeyDown}{}

\item{parentActive}{Logical. }
}
