% Auto-generated: do not edit by hand
\name{''MicrophoneDashUiComponents}

\alias{''MicrophoneDashUiComponents}

\title{MicrophoneDashUiComponents component}

\description{

}

\usage{
''MicrophoneDashUiComponents(id=NULL, backgroundColor=NULL, strokeColor=NULL, className=NULL,
audioBitsPerSecond=NULL, mimeType=NULL, height=NULL, record=NULL, onStop=NULL,
onData=NULL, realData=NULL, recordedBlob=NULL, visualSetting=NULL, width=NULL)
}

\arguments{
\item{id}{Character. }

\item{backgroundColor}{Character. }

\item{strokeColor}{Character. }

\item{className}{Character. }

\item{audioBitsPerSecond}{Numeric. }

\item{mimeType}{Character. }

\item{height}{Numeric. }

\item{record}{Logical. }

\item{onStop}{}

\item{onData}{}

\item{realData}{Character. }

\item{recordedBlob}{Named list. }

\item{visualSetting}{Character. }

\item{width}{Numeric. }
}
