% Auto-generated: do not edit by hand
\name{''PageLinks}

\alias{''PageLinks}

\title{PageLinks component}

\description{

}

\usage{
''PageLinks(id=NULL, value=NULL, page=NULL, onClick=NULL)
}

\arguments{
\item{id}{Character. }

\item{value}{Unnamed list. }

\item{page}{Numeric. }

\item{onClick}{}
}
