% Auto-generated: do not edit by hand
\name{''PaginatorDashUiComponents}

\alias{''PaginatorDashUiComponents}

\title{PaginatorDashUiComponents component}

\description{

}

\usage{
''PaginatorDashUiComponents(id=NULL, totalRecords=NULL, rows=NULL, first=NULL, pageLinkSize=NULL,
rowsPerPageOptions=NULL, style=NULL, className=NULL, template=NULL,
onPageChange=NULL, leftContent=NULL, rightContent=NULL,
currentPageReportTemplate=NULL, alwaysShow=NULL)
}

\arguments{
\item{id}{Character. }

\item{totalRecords}{Numeric. }

\item{rows}{Numeric. }

\item{first}{Numeric. }

\item{pageLinkSize}{Numeric. }

\item{rowsPerPageOptions}{Unnamed list. }

\item{style}{Named list. }

\item{className}{Character. }

\item{template}{Character. }

\item{onPageChange}{}

\item{leftContent}{Logical | numeric | character | named list | unnamed list. }

\item{rightContent}{Logical | numeric | character | named list | unnamed list. }

\item{currentPageReportTemplate}{Logical | numeric | character | named list | unnamed list. }

\item{alwaysShow}{Logical. }
}
