% Auto-generated: do not edit by hand
\name{''PanelMenu}

\alias{''PanelMenu}

\title{PanelMenu component}

\description{
PanelMenu
}

\usage{
''PanelMenu(id=NULL, model=NULL, style=NULL, className=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}
}
