% Auto-generated: do not edit by hand
\name{''PanelMenuComponent}

\alias{''PanelMenuComponent}

\title{PanelMenuComponent component}

\description{
PanelMenuComponent
}

\usage{
''PanelMenuComponent(id=NULL, model=NULL, style=NULL, className=NULL, panelmenuItem=NULL,
panelsubmenu=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{panelmenuItem}{Named list. Current menuitem}

\item{panelsubmenu}{Named list. Selected submenu item}
}
