% Auto-generated: do not edit by hand
\name{''PDFViewerComponent}

\alias{''PDFViewerComponent}

\title{PDFViewerComponent component}

\description{
PDF Viewer component with controls Page navigation Zoom Rotation
}

\usage{
''PDFViewerComponent(url=NULL, date64=NULL)
}

\arguments{
\item{url}{Character. The url of the PDF document}

\item{date64}{Character. PDF file encoded in base64}
}
