% Auto-generated: do not edit by hand
\name{''StepsComponent}

\alias{''StepsComponent}

\title{StepsComponent component}

\description{
WrapperComponent on the Steps component Adds callback
}

\usage{
''StepsComponent(id=NULL, model=NULL, activeIndex=NULL, readOnly=NULL, style=NULL,
className=NULL, onSelect=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{activeIndex}{Numeric. Index of the active item.}

\item{readOnly}{Logical. Whether the items are clickable or not.}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{onSelect}{Callback when item is selected}
}
