% Auto-generated: do not edit by hand
\name{''TrelloComponent}

\alias{''TrelloComponent}

\title{TrelloComponent component}

\description{
Component to render a button element
}

\usage{
''TrelloComponent(value=NULL, draggable=NULL, laneDraggable=NULL, cardDraggable=NULL,
collapsibleLanes=NULL, editable=NULL, canAddLanes=NULL, hideCardDeleteIcon=NULL,
editLaneTitle=NULL)
}

\arguments{
\item{value}{Named list. }

\item{draggable}{Logical. }

\item{laneDraggable}{Logical. }

\item{cardDraggable}{Logical. }

\item{collapsibleLanes}{Logical. }

\item{editable}{Logical. }

\item{canAddLanes}{Logical. }

\item{hideCardDeleteIcon}{Logical. }

\item{editLaneTitle}{Logical. }
}
