% Auto-generated: do not edit by hand
\name{breadcrumbsDashUiComponents}

\alias{breadcrumbsDashUiComponents}

\title{BreadcrumbsDashUiComponents component}

\description{

}

\usage{
breadcrumbsDashUiComponents(id=NULL, model=NULL, home=NULL, style=NULL, className=NULL)
}

\arguments{
\item{id}{Character. }

\item{model}{Unnamed list. }

\item{home}{Logical | numeric | character | named list | unnamed list. }

\item{style}{Named list. }

\item{className}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

