% Auto-generated: do not edit by hand
\name{button}

\alias{button}

\title{Button component}

\description{
Component to render a button element
}

\usage{
button(label=NULL, icon=NULL, iconPos=NULL, tooltip=NULL,
tooltipOptions=NULL)
}

\arguments{
\item{label}{Character. Text of the button.}

\item{icon}{Character. Name of the icon.}

\item{iconPos}{Character. Position of the icon, valid values are "left" and "right".}

\item{tooltip}{Character. Content of the tooltip.}

\item{tooltipOptions}{Named list. Configuration of the tooltip, refer to the tooltip documentation for more information.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

