% Auto-generated: do not edit by hand
\name{currentPageReport}

\alias{currentPageReport}

\title{CurrentPageReport component}

\description{

}

\usage{
currentPageReport(id=NULL, pageCount=NULL, page=NULL, template=NULL)
}

\arguments{
\item{id}{Character. }

\item{pageCount}{Numeric. }

\item{page}{Numeric. }

\item{template}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

