% Auto-generated: do not edit by hand
\name{geolocatorComponent}

\alias{geolocatorComponent}

\title{GeolocatorComponent component}

\description{

}

\usage{
geolocatorComponent(id=NULL, positionOptions=NULL, watchPosition=NULL, userDecisionTimeout=NULL,
suppressLocationOnMount=NULL, isOptimisticGeolocationEnabled=NULL,
errorMsg=NULL, onSuccess=NULL, onError=NULL, coords=NULL)
}

\arguments{
\item{id}{Character. }

\item{positionOptions}{Named list. }

\item{watchPosition}{Logical. }

\item{userDecisionTimeout}{Numeric. }

\item{suppressLocationOnMount}{Logical. }

\item{isOptimisticGeolocationEnabled}{Logical. }

\item{errorMsg}{Named list. }

\item{onSuccess}{}

\item{onError}{}

\item{coords}{Named list. }
}
