% Auto-generated: do not edit by hand
\name{geolocatorDashUiComponents}

\alias{geolocatorDashUiComponents}

\title{GeolocatorDashUiComponents component}

\description{

}

\usage{
geolocatorDashUiComponents(id=NULL)
}

\arguments{
\item{id}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

