% Auto-generated: do not edit by hand
\name{idleTimer}

\alias{idleTimer}

\title{IdleTimer component}

\description{

}

\usage{
idleTimer(id=NULL, action=NULL, active=NULL, idle=NULL, timeout=NULL,
debounce=NULL, startOnMount=NULL, stopOnIdle=NULL,
passive=NULL, capture=NULL)
}

\arguments{
\item{id}{Character. }

\item{action}{Named list. }

\item{active}{Named list. }

\item{idle}{Named list. }

\item{timeout}{Numeric. }

\item{debounce}{Numeric. }

\item{startOnMount}{Logical. }

\item{stopOnIdle}{Logical. }

\item{passive}{Logical. }

\item{capture}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

