% Auto-generated: do not edit by hand
\name{idleTimerComponent}

\alias{idleTimerComponent}

\title{IdleTimerComponent component}

\description{

}

\usage{
idleTimerComponent(id=NULL, timeout=NULL, action=NULL, idle=NULL, active=NULL,
acType=NULL, xPosition=NULL, yPosition=NULL, movementX=NULL,
movementY=NULL, pageX=NULL, pageY=NULL, screenX=NULL,
screenY=NULL, timeStamp=NULL, type=NULL, code=NULL,
key=NULL, which=NULL)
}

\arguments{
\item{id}{Character. }

\item{timeout}{Numeric. }

\item{action}{Named list. }

\item{idle}{Character. }

\item{active}{Named list. }

\item{acType}{Character. }

\item{xPosition}{Numeric. }

\item{yPosition}{Numeric. }

\item{movementX}{Numeric. }

\item{movementY}{Numeric. }

\item{pageX}{Numeric. }

\item{pageY}{Numeric. }

\item{screenX}{Numeric. }

\item{screenY}{Numeric. }

\item{timeStamp}{Numeric. }

\item{type}{Character. }

\item{code}{Character. }

\item{key}{Character. }

\item{which}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

