% Auto-generated: do not edit by hand
\name{inputText}

\alias{inputText}

\title{InputText component}

\description{
InputText Bar. displays on top of the search box component if filter is set to true
}

\usage{
inputText(onInput=NULL, onKeyPress=NULL, keyfilter=NULL, validateOnly=NULL, tooltip=NULL,
tooltipOptions=NULL)
}

\arguments{
\item{onInput}{Callback to invoke when input is in focus changes.}

\item{onKeyPress}{Callback to invoke when value of input changes.}

\item{keyfilter}{Logical | numeric | character | named list | unnamed list. Filter value}

\item{validateOnly}{Logical. Boolean}

\item{tooltip}{Character. Tooltip text}

\item{tooltipOptions}{Named list. Tooltip options. An object}
}
