% Auto-generated: do not edit by hand
\name{listBoxHeader}

\alias{listBoxHeader}

\title{ListBoxHeader component}

\description{
Header component of the listbox
}

\usage{
listBoxHeader(filter=NULL, disabled=NULL, onFilter=NULL)
}

\arguments{
\item{filter}{Character. When specified, displays a filter input at header.}

\item{disabled}{Logical. When specified, disables the component.}

\item{onFilter}{Callback to invoke when value of filter list changes}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

