% Auto-generated: do not edit by hand
\name{menuBar}

\alias{menuBar}

\title{MenuBar component}

\description{
MenuBar
}

\usage{
menuBar(id=NULL, model=NULL, style=NULL, className=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

