% Auto-generated: do not edit by hand
\name{menubarSub}

\alias{menubarSub}

\title{MenubarSub component}

\description{

}

\usage{
menubarSub(model=NULL, root=NULL, className=NULL, popup=NULL,
onLeafClick=NULL, onKeyDown=NULL, parentActive=NULL)
}

\arguments{
\item{model}{Logical | numeric | character | named list | unnamed list. }

\item{root}{Logical. }

\item{className}{Character. }

\item{popup}{Logical. }

\item{onLeafClick}{}

\item{onKeyDown}{}

\item{parentActive}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

