% Auto-generated: do not edit by hand
\name{microphoneDashUiComponents}

\alias{microphoneDashUiComponents}

\title{MicrophoneDashUiComponents component}

\description{

}

\usage{
microphoneDashUiComponents(id=NULL, backgroundColor=NULL, strokeColor=NULL, className=NULL,
audioBitsPerSecond=NULL, mimeType=NULL, height=NULL, record=NULL, realData=NULL,
recordedBlob=NULL, visualSetting=NULL, width=NULL)
}

\arguments{
\item{id}{Character. }

\item{backgroundColor}{Character. }

\item{strokeColor}{Character. }

\item{className}{Character. }

\item{audioBitsPerSecond}{Numeric. }

\item{mimeType}{Character. }

\item{height}{Numeric. }

\item{record}{Logical. }

\item{realData}{Character. }

\item{recordedBlob}{Named list. }

\item{visualSetting}{Character. }

\item{width}{Numeric. }
}
