% Auto-generated: do not edit by hand
\name{nextPageLink}

\alias{nextPageLink}

\title{NextPageLink component}

\description{

}

\usage{
nextPageLink(id=NULL, disabled=NULL, onClick=NULL)
}

\arguments{
\item{id}{Character. }

\item{disabled}{Logical. }

\item{onClick}{}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

