% Auto-generated: do not edit by hand
\name{organizationChartComponent}

\alias{organizationChartComponent}

\title{OrganizationChartComponent component}

\description{

}

\usage{
organizationChartComponent(id=NULL, value=NULL, selectionMode=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Unnamed list. An array of nested TreeNodes}

\item{selectionMode}{Character. Defines the selection mode, valid values "single" and "multiple".}
}
