% Auto-generated: do not edit by hand
\name{pageLinks}

\alias{pageLinks}

\title{PageLinks component}

\description{

}

\usage{
pageLinks(id=NULL, value=NULL, page=NULL, onClick=NULL)
}

\arguments{
\item{id}{Character. }

\item{value}{Unnamed list. }

\item{page}{Numeric. }

\item{onClick}{}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

