% Auto-generated: do not edit by hand
\name{panelMenuComponent}

\alias{panelMenuComponent}

\title{PanelMenuComponent component}

\description{
PanelMenuComponent
}

\usage{
panelMenuComponent(id=NULL, model=NULL, style=NULL, className=NULL, menuitem=NULL, submenu=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{menuitem}{Named list. Current menuitem}

\item{submenu}{Named list. Selected submenu item}
}
