% Auto-generated: do not edit by hand
\name{pMSChildrenComponent}

\alias{pMSChildrenComponent}

\title{PMSChildrenComponent component}

\description{

}

\usage{
pMSChildrenComponent(data=NULL, girlDataUpdate=NULL, range=NULL, delete=NULL)
}

\arguments{
\item{data}{Named list. }

\item{girlDataUpdate}{}

\item{range}{Numeric. }

\item{delete}{}
}
