% Auto-generated: do not edit by hand
\name{pMSComponent}

\alias{pMSComponent}

\title{PMSComponent component}

\description{

}

\usage{
pMSComponent(value=NULL, id=NULL, newGirl=NULL, updateData=NULL)
}

\arguments{
\item{value}{Unnamed list. }

\item{id}{Character. }

\item{newGirl}{Named list. }

\item{updateData}{Named list. }
}
