% Auto-generated: do not edit by hand
\name{prevPageLink}

\alias{prevPageLink}

\title{PrevPageLink component}

\description{

}

\usage{
prevPageLink(id=NULL, disabled=NULL, onClick=NULL)
}

\arguments{
\item{id}{Character. }

\item{disabled}{Logical. }

\item{onClick}{}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

