% Auto-generated: do not edit by hand
\name{ratingComponent}

\alias{ratingComponent}

\title{RatingComponent component}

\description{
Rating component
}

\usage{
ratingComponent(id=NULL, value=NULL, disabled=NULL, readonly=NULL, stars=NULL, cancel=NULL,
style=NULL, className=NULL, tooltip=NULL, tooltipOptions=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{value}{Numeric. Value of the rating.}

\item{disabled}{Logical. When present, it specifies that the element should be disabled.}

\item{readonly}{Logical. When present, changing the value is not possible.}

\item{stars}{Numeric. Number of stars.}

\item{cancel}{Logical. When specified a cancel icon is displayed to allow removing the value.}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. ClassName of the component.}

\item{tooltip}{Character. Content of the tooltip.}

\item{tooltipOptions}{Named list. Configuration of the tooltip, refer to the tooltip documentation for more information.}
}
