% Auto-generated: do not edit by hand
\name{steps}

\alias{steps}

\title{Steps component}

\description{
Steps
}

\usage{
steps(id=NULL, model=NULL, activeIndex=NULL, readOnly=NULL, style=NULL,
className=NULL, onSelect=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{activeIndex}{Numeric. Index of the active item.}

\item{readOnly}{Logical. Whether the items are clickable or not.}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{onSelect}{Callback to invoke when the new step is selected.}
}
