% Auto-generated: do not edit by hand
\name{stepsComponent}

\alias{stepsComponent}

\title{StepsComponent component}

\description{
WrapperComponent on the Steps component Adds callback
}

\usage{
stepsComponent(id=NULL, model=NULL, activeIndex=NULL, readOnly=NULL,
style=NULL, className=NULL, onSelect=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{activeIndex}{Numeric. Index of the active item.}

\item{readOnly}{Logical. Whether the items are clickable or not.}

\item{style}{Named list. Inline style of the component.}

\item{className}{Character. Style class of the component.}

\item{onSelect}{Callback when item is selected}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

