% Auto-generated: do not edit by hand
\name{treeDashUiComponents}

\alias{treeDashUiComponents}

\title{TreeDashUiComponents component}

\description{

}

\usage{
treeDashUiComponents(id=NULL, value=NULL, disabled=NULL, selectionMode=NULL, selectionKeys=NULL,
onSelectionChange=NULL, contextMenuSelectionKey=NULL,
onContextMenuSelectionChange=NULL, expandedKeys=NULL, style=NULL,
className=NULL, contentStyle=NULL, contentClassName=NULL, metaKeySelection=NULL,
propagateSelectionUp=NULL, propagateSelectionDown=NULL, loading=NULL,
loadingIcon=NULL, dragdropScope=NULL, filter=NULL, filterBy=NULL,
filterMode=NULL, filterPlaceholder=NULL, nodeTemplate=NULL, onSelect=NULL,
onUnselect=NULL, onExpand=NULL, onCollapse=NULL, onToggle=NULL, onDragDrop=NULL,
onContextMenu=NULL)
}

\arguments{
\item{id}{Character. }

\item{value}{Logical | numeric | character | named list | unnamed list. }

\item{disabled}{Logical. }

\item{selectionMode}{Character. }

\item{selectionKeys}{Logical | numeric | character | named list | unnamed list. }

\item{onSelectionChange}{}

\item{contextMenuSelectionKey}{Logical | numeric | character | named list | unnamed list. }

\item{onContextMenuSelectionChange}{}

\item{expandedKeys}{Named list. }

\item{style}{Named list. }

\item{className}{Character. }

\item{contentStyle}{Named list. }

\item{contentClassName}{Character. }

\item{metaKeySelection}{Logical. }

\item{propagateSelectionUp}{Logical. }

\item{propagateSelectionDown}{Logical. }

\item{loading}{Logical. }

\item{loadingIcon}{Character. }

\item{dragdropScope}{Character. }

\item{filter}{Logical. }

\item{filterBy}{Logical | numeric | character | named list | unnamed list. }

\item{filterMode}{Character. }

\item{filterPlaceholder}{Character. }

\item{nodeTemplate}{}

\item{onSelect}{}

\item{onUnselect}{}

\item{onExpand}{}

\item{onCollapse}{}

\item{onToggle}{}

\item{onDragDrop}{}

\item{onContextMenu}{}
}
