% Auto-generated: do not edit by hand
\name{trelloComponent}

\alias{trelloComponent}

\title{TrelloComponent component}

\description{
Component to render a button element
}

\usage{
trelloComponent(draggable=NULL, laneDraggable=NULL, cardDraggable=NULL, collapsibleLanes=NULL,
editable=NULL, canAddLanes=NULL, hideCardDeleteIcon=NULL, editLaneTitle=NULL)
}

\arguments{
\item{draggable}{Logical. }

\item{laneDraggable}{Logical. }

\item{cardDraggable}{Logical. }

\item{collapsibleLanes}{Logical. }

\item{editable}{Logical. }

\item{canAddLanes}{Logical. }

\item{hideCardDeleteIcon}{Logical. }

\item{editLaneTitle}{Logical. }
}
