% Auto-generated: do not edit by hand
\name{uITreeNode}

\alias{uITreeNode}

\title{UITreeNode component}

\description{

}

\usage{
uITreeNode(node=NULL, index=NULL, last=NULL, parent=NULL, path=NULL,
disabled=NULL, selectionMode=NULL, selectionKeys=NULL,
contextMenuSelectionKey=NULL, metaKeySelection=NULL,
expandedKeys=NULL, propagateSelectionUp=NULL,
propagateSelectionDown=NULL, dragdropScope=NULL,
ariaLabel=NULL, ariaLabelledBy=NULL, nodeTemplate=NULL,
isNodeLeaf=NULL, onSelect=NULL, onUnselect=NULL,
onExpand=NULL, onCollapse=NULL, onToggle=NULL,
onSelectionChange=NULL, onContextMenuSelectionChange=NULL,
onPropagateUp=NULL, onDragStart=NULL, onDragEnd=NULL,
onDrop=NULL, onDropPoint=NULL, onContextMenu=NULL)
}

\arguments{
\item{node}{Named list. }

\item{index}{Numeric. }

\item{last}{Logical. }

\item{parent}{Named list. }

\item{path}{Character. }

\item{disabled}{Logical. }

\item{selectionMode}{Character. }

\item{selectionKeys}{Logical | numeric | character | named list | unnamed list. }

\item{contextMenuSelectionKey}{Logical | numeric | character | named list | unnamed list. }

\item{metaKeySelection}{Logical. }

\item{expandedKeys}{Named list. }

\item{propagateSelectionUp}{Logical. }

\item{propagateSelectionDown}{Logical. }

\item{dragdropScope}{Character. }

\item{ariaLabel}{Character. }

\item{ariaLabelledBy}{Character. }

\item{nodeTemplate}{}

\item{isNodeLeaf}{}

\item{onSelect}{}

\item{onUnselect}{}

\item{onExpand}{}

\item{onCollapse}{}

\item{onToggle}{}

\item{onSelectionChange}{}

\item{onContextMenuSelectionChange}{}

\item{onPropagateUp}{}

\item{onDragStart}{}

\item{onDragEnd}{}

\item{onDrop}{}

\item{onDropPoint}{}

\item{onContextMenu}{}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

