% Auto-generated: do not edit by hand
\name{''GeolocatorComponent}

\alias{''GeolocatorComponent}

\title{GeolocatorComponent component}

\description{

}

\usage{
''GeolocatorComponent(position=NULL, id=NULL, positionOptions=NULL, watchPosition=NULL,
userDecisionTimeout=NULL, suppressLocationOnMount=NULL,
isOptimisticGeolocationEnabled=NULL, errorMsg=NULL, onSuccess=NULL, onError=NULL)
}

\arguments{
\item{position}{Named list. }

\item{id}{Character. }

\item{positionOptions}{Named list. }

\item{watchPosition}{Logical. }

\item{userDecisionTimeout}{Numeric. }

\item{suppressLocationOnMount}{Logical. }

\item{isOptimisticGeolocationEnabled}{Logical. }

\item{errorMsg}{Named list. }

\item{onSuccess}{}

\item{onError}{}
}
