% Auto-generated: do not edit by hand
\name{gPayButtonComponent}

\alias{gPayButtonComponent}

\title{GPayButtonComponent component}

\description{

}

\usage{
gPayButtonComponent(id=NULL, totalPrice=NULL, totalPriceStatus=NULL, currencyCode=NULL,
countryCode=NULL, merchantInfo=NULL, paymentMethods=NULL)
}

\arguments{
\item{id}{Character. }

\item{totalPrice}{Character. }

\item{totalPriceStatus}{Character. }

\item{currencyCode}{Character. }

\item{countryCode}{Character. }

\item{merchantInfo}{Named list. }

\item{paymentMethods}{Unnamed list. }
}
