% Auto-generated: do not edit by hand
\name{menuBarComponent}

\alias{menuBarComponent}

\title{MenuBarComponent component}

\description{
MenuBarComponent
}

\usage{
menuBarComponent(id=NULL, model=NULL, style=NULL, inputPlaceholder=NULL, btnicon=NULL,
btnlabel=NULL, className=NULL, menubarItem=NULL, submenubarItem=NULL)
}

\arguments{
\item{id}{Character. Unique identifier of the element.}

\item{model}{Unnamed list. An array of menuitems.}

\item{style}{Named list. Inline style of the component.}

\item{inputPlaceholder}{Character. Input placeholder}

\item{btnicon}{Character. Button Icon}

\item{btnlabel}{Character. Button label}

\item{className}{Character. Style class of the component.}

\item{menubarItem}{Named list. Current Menu object}

\item{submenubarItem}{Named list. Currently selected item}
}
