% Auto-generated: do not edit by hand
\name{trelloComponent}

\alias{trelloComponent}

\title{TrelloComponent component}

\description{

}

\usage{
trelloComponent(id=NULL, value=NULL, draggable=NULL, laneDraggable=NULL,
cardDraggable=NULL, collapsibleLanes=NULL, editable=NULL,
canAddLanes=NULL, hideCardDeleteIcon=NULL,
editLaneTitle=NULL, changedData=NULL, cardId=NULL,
landId=NULL, event=NULL, sourceLaneId=NULL,
targetLaneId=NULL, position=NULL, cardDetails=NULL,
landParams=NULL, metaData=NULL, card=NULL)
}

\arguments{
\item{id}{Character. }

\item{value}{Named list. }

\item{draggable}{Logical. }

\item{laneDraggable}{Logical. }

\item{cardDraggable}{Logical. }

\item{collapsibleLanes}{Logical. }

\item{editable}{Logical. }

\item{canAddLanes}{Logical. }

\item{hideCardDeleteIcon}{Logical. }

\item{editLaneTitle}{Logical. }

\item{changedData}{Named list. }

\item{cardId}{Character. }

\item{landId}{Character. }

\item{event}{Character. }

\item{sourceLaneId}{Character. }

\item{targetLaneId}{Character. }

\item{position}{Numeric. }

\item{cardDetails}{Named list. }

\item{landParams}{Named list. }

\item{metaData}{Named list. }

\item{card}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

